requirements.txt
Django==4.2.7
django-bootstrap5==23.3
python-docx==1.1.0
scikit-learn==1.3.2
pillow==10.1.0
gunicorn==21.2.0